Separate install / access of fie to different files to avoid circular dependencies (#23531)
* Separate install / access of fie to different files to avoid circular dependencies.

* Move getFie() to iframe-helper.js

* move

* lint
