+++ src/friendly-iframe-embed.js
@@ -15,13 +15,13 @@
  */
 
 import {CommonSignals} from './common-signals';
+import {FIE_EMBED_PROP} from './iframe-helper';
 import {LEGACY_ELEMENTS, stubLegacyElements} from './service/extensions-impl';
 import {Observable} from './observable';
 import {Services} from './services';
 import {Signals} from './utils/signals';
 import {cssText as ampDocCss} from '../build/ampdoc.css';
 import {cssText as ampSharedCss} from '../build/ampshared.css';
-import {closestAncestorElementBySelector, escapeHtml} from './dom';
 import {
   copyElementToChildWindow,
   stubElementIfNotKnown,
@@ -35,6 +35,7 @@ import {
   installServiceInEmbedIfEmbeddable,
   setParentWindow,
 } from './service';
+import {escapeHtml} from './dom';
 import {getExperimentBranch, isExperimentOn} from './experiments';
 import {getMode} from './mode';
 import {installAmpdocServices} from './service/core-services';
@@ -45,7 +46,6 @@ import {installCustomElements as installRegisterElement} from 'document-register
 import {installStylesForDoc, installStylesLegacy} from './style-installer';
 import {installTimerInEmbedWindow} from './service/timer-impl';
 import {isDocumentReady} from './document-ready';
-
 import {layoutRectLtwh, moveLayoutRect} from './layout-rect';
 import {loadPromise} from './event-helper';
 import {
@@ -56,9 +56,6 @@ import {
   setStyles,
 } from './style';
 import {toWin} from './types';
-
-/** @const {string} */
-const EMBED_PROP = '__AMP_EMBED__';
 
 /** @const {!Array<string>} */
 const EXCLUDE_INI_LOAD = [
@@ -131,18 +128,6 @@ function isSrcdocSupported() {
  */
 export function setFriendlyIframeEmbedVisible(embed, visible) {
   embed.setVisible_(visible);
-}
-
-/**
- * Returns the embed created using `installFriendlyIframeEmbed` or `null`.
- * Caution: This will only return the FIE after the iframe has 'loaded'. If you
- * are checking before this signal you may be in a race condition that returns
- * null.
- * @param {!HTMLIFrameElement} iframe
- * @return {?FriendlyIframeEmbed}
- */
-export function getFriendlyIframeEmbedOptional(iframe) {
-  return /** @type {?FriendlyIframeEmbed} */ (iframe[EMBED_PROP]);
 }
 
 /**
@@ -256,7 +241,7 @@ export function installFriendlyIframeEmbed(
         ? ampdocService.installFieDoc(spec.url, childWin, {signals})
         : null;
     const embed = new FriendlyIframeEmbed(iframe, spec, loadedPromise, ampdoc);
-    iframe[EMBED_PROP] = embed;
+    iframe[FIE_EMBED_PROP] = embed;
 
     // Add extensions.
     if (ampdoc && ampdocFieExperimentOn) {
@@ -872,17 +857,6 @@ export function whenContentIniLoad(elementOrAmpDoc, hostWin, rect) {
       });
       return Promise.all(promises);
     });
-}
-
-/**
- * @param {!Element} element
- * @return {boolean}
- */
-export function isInFie(element) {
-  return (
-    element.classList.contains('i-amphtml-fie') ||
-    !!closestAncestorElementBySelector(element, '.i-amphtml-fie')
-  );
 }
 
 /**
